Improve output when fetching input
fetch_inputs.py now prints out whether each day was skipped, cached, or
downloaded. In addition, it will print a warning if all days are skipped
since that means that the user might be running it in the wrong
directory. This will allow the user to better understand what is
happening.